# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Woodkraft API",
        "description": "API pour le site e-commerce de mobilier en bois. Gestion des produits, catégories, variations et images.",
        "contact": {
            "email": "contact@woodkraft.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8888",
            "description": "Serveur de développement MAMP"
        }
    ],
    "paths": {
        "/api/v1/auth/register": {
            "post": {
                "tags": [
                    "Auth"
                ],
                "summary": "Créer un compte client",
                "description": "Inscription d'un nouveau client",
                "operationId": "ff564de52b6fcd86070c75342a9d31f0",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "type": "string",
                                        "example": "Jean Dupont"
                                    },
                                    "email": {
                                        "type": "string",
                                        "format": "email",
                                        "example": "jean@example.com"
                                    },
                                    "password": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    },
                                    "password_confirmation": {
                                        "type": "string",
                                        "format": "password",
                                        "example": "password"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Compte créé"
                    },
                    "422": {
                        "description": "Données invalides"
                    }
                }
            }
        },
        "/api/customer/cart": {
            "get": {
                "tags": [
                    "Cart"
                ],
                "summary": "Récupérer le panier du client connecté",
                "description": "Get user's cart items",
                "operationId": "fbc5c2a8e7c0fb06aedfdf2ef2219200",
                "responses": {
                    "200": {
                        "description": "Articles du panier"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Cart"
                ],
                "summary": "Vider le panier",
                "description": "Clear cart",
                "operationId": "f77f1e3ad9d80243d30c412a0fa70d34",
                "responses": {
                    "200": {
                        "description": "Panier vidé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/api/customer/cart/sync": {
            "post": {
                "tags": [
                    "Cart"
                ],
                "summary": "Synchroniser le panier complet",
                "description": "Sync full cart from frontend",
                "operationId": "1769be02b7cdadd316229a63efb0a09f",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "productId": {
                                                    "type": "integer"
                                                },
                                                "quantity": {
                                                    "type": "integer"
                                                },
                                                "selectedColor": {
                                                    "type": "object",
                                                    "nullable": true
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Panier synchronisé"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
...
}